Bug 1627271 - Order 'Similar Jobs' first by push id, then by task's start time. (#6266)
The most common issue with failed tasks except intermittent failures are
check-in related issues. Until now, the tasks in 'Similar Jobs' were ordered by
the task's start time which was confusing if tasks had been requested to run
again ("retrigger") because the list jumped between pushes. With the new order
by pushes, issues can be more easily associated with a push/check-in.

The downside is that infrastructure related issues are harder to spot but in
general these don't affect only one kind of task (which the 'Similar Jobs' tab
focuses on).